오류수정 코드 간결화 할 것

+++ src/main/java/bit38_7/MapConvertor/repository/file/FileRepository.java
@@ -11,7 +11,7 @@ public interface FileRepository {
 
 	void saveFloor(int buildingId, int floorNum, byte[] floorData);
 
-	List<BuildingResponse> userBuildingList(int userId);
+	List<BuildingResponse> userBuildingList(Long userId);
 
 	byte[] findBuildingFile(int buildingId);
